Add plugin for improved git commits in VIM
It's recommended that the body of the comment has a column width of no
greater than 72, so I knew the 'colorcolumn' feature of VIM would come
in handy.

I needed to have this set automatically, though, based on the line
number in which I'm editing, since I also added the comments.

If this doesn't go to plan, and messes up what is actually displayed on
GitHub itself, I do apologise.