Make more robust and enforce rules
The subject line (#2) now has a hard limit of 50 columns, instead of
just relying on the user to adhere to it.

If you exceed the subject column limit, you'll lose everything after
that, but it should be as accessible as anything else once deleted.

Automatic wrapping is now done during insert mode, when the user is in
the 'body' of the commit message (this area). This just makes it easier
for the user to quickly produce text, without needing to waste time
formatting it.

The spelling is now checked, since it's probably, mostly going to be a
traditional language; in other words, not code.