Persistent undo is a life-saver
For me, the killer feature of Vim is the undo system. Every time I try to switch
editors I miss persistent undo. This allows me to undo changes even between vim
sessions. This feature works beautifully with the other feature of vim that I
should use but don't. The undo list/undo tree that allows you to see the entire
history of the file, assuming you have the undo history.